requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --generate-hashes --no-emit-index-url --resolver=backtracking --strip-extras requirements.in
#
attrs==22.2.0 \
    --hash=sha256:29e95c7f6778868dbd49170f98f8818f78f3dc5e0e37c0b1f474e3561b240836 \
    --hash=sha256:c9227bfc2f01993c03f68db37d1d15c9690188323c067c641f1a35ca58185f99
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.136 \
    --hash=sha256:90607fe6563a4ea71ea596d31c27f96671344eef8e1a752fa506da1d7655c062 \
    --hash=sha256:a7f1bb67860505ed9894e45e6d0427ef6d0bec611d620f024ee66dbf0ac39369
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1 \
    --hash=sha256:9834cdb150c5590aea4e5eba6de2a89b4c60617451181c524810c5a75154565c \
    --hash=sha256:a2fad1a5a35a94a465efe60859f91e45dacc33261fb9bbf1cf9bbc6e2f70e9d6
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.112 \
    --hash=sha256:cb965e0e2566db5b6dc3c5be59705e95a0e88a0cafc116f904e3144012ebfee0 \
    --hash=sha256:e07b5d79fc0282956dd3a0c3d396f7d0007349e6ee89815158a521adb4a2e43f
    # via aws-cdk-lib
aws-cdk-lib==2.73.0 \
    --hash=sha256:60271826f4d53267b39c43aaba93eecbcd2b85c1a1ae206ce56f347344903554 \
    --hash=sha256:9e93044d19ae26ef4303b39cbd4b083f8b183bb9211bc2f9b76698a8c765d8ad
    # via -r requirements.in
cattrs==22.2.0 \
    --hash=sha256:bc12b1f0d000b9f9bee83335887d532a1d3e99a833d1bf0882151c97d3e68c21 \
    --hash=sha256:f0eed5642399423cf656e7b66ce92cdc5b963ecafd041d1b24d136fdde7acf6d
    # via jsii
constructs==10.1.307 \
    --hash=sha256:2b8d90e09d50f81d4ba1a451eb1ce3480b8ba3fd238396c21652cf2306bdba02 \
    --hash=sha256:870b94d8440cc38aff3490c45c7bb99f1fc32a2756aa09dbaa4db5cb517d717a
    # via
    #   -r requirements.in
    #   aws-cdk-lib
exceptiongroup==1.1.1 \
    --hash=sha256:232c37c63e4f682982c8b6459f33a8981039e5fb8756b2074364e5055c498c9e \
    --hash=sha256:d484c3090ba2889ae2928419117447a14daf3c1231d5e30d0aae34f354f01785
    # via cattrs
importlib-resources==5.12.0 \
    --hash=sha256:4be82589bf5c1d7999aedf2a45159d10cb3ca4f19b2271f8792bc8e6da7b22f6 \
    --hash=sha256:7b1deeebbf351c7578e09bf2f63fa2ce8b5ffec296e0d349139d43cca061a81a
    # via jsii
jsii==1.80.0 \
    --hash=sha256:4da63ab99f2696cd063574460c94221f0a7de9d345e71dfb19dfbcecf8ca8355 \
    --hash=sha256:ea3cace063f6a47cdf0a74c929618d779efab426fedb7692a8ac1b9b29797f8c
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
publication==0.0.3 \
    --hash=sha256:0248885351febc11d8a1098d5c8e3ab2dabcf3e8c0c96db1e17ecd12b53afbe6 \
    --hash=sha256:68416a0de76dddcdd2930d1c8ef853a743cc96c82416c4e4d3b5d901c6276dc4
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.8.2 \
    --hash=sha256:0123cacc1627ae19ddf3c27a5de5bd67ee4586fbdd6440d9748f8abb483d3e86 \
    --hash=sha256:961d03dc3453ebbc59dbdea9e4e11c5651520a876d0f4db161e8674aae935da9
    # via jsii
six==1.16.0 \
    --hash=sha256:1e61c37477a1626458e36f7b1d82aa5c9b094fa4802892072e49de9c60c4c926 \
    --hash=sha256:8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254
    # via python-dateutil
typeguard==2.13.3 \
    --hash=sha256:00edaa8da3a133674796cf5ea87d9f4b4c367d77476e185e80251cc13dfbb8c4 \
    --hash=sha256:5e3e3be01e887e7eafae5af63d1f36c849aaa94e3a0112097312aabfa16284f1
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.5.0 \
    --hash=sha256:5cb5f4a79139d699607b3ef622a1dedafa84e115ab0024e0d9c044a9479ca7cb \
    --hash=sha256:fb33085c39dd998ac16d1431ebc293a8b3eedd00fd4a32de0ff79002c19511b4
    # via jsii
